+++ library/register/styles-scripts.php
@@ -15,6 +15,7 @@ if(!function_exists('pix_mitarbeiter_register_files')):
 		if ( get_option('include_css')){
 			wp_register_style( 'mitarbeiter', HTTP_PIX_MITARBEITER_PATH.'/assets/css/mitarbeiter.css');
 			wp_enqueue_style( 'mitarbeiter' );
+			
 		}
 		
 		if ( get_option('include_accordeon')){
@@ -24,7 +25,9 @@ if(!function_exists('pix_mitarbeiter_register_files')):
 			wp_register_script( 'mitarbeiter', HTTP_PIX_MITARBEITER_PATH.'/assets/js/mitarbeiter.js');
 			wp_enqueue_script( 'mitarbeiter' );
 			
-			
+			//Load Dashicons
+		    wp_enqueue_style('dashicons');
+					
 		}
 		
 	}