+++ CHANGES.md
@@ -3,6 +3,7 @@
 #### 0.1.6 (21/05/14)
 
 * added filtered user-output via add_filter
+* added WP dashicons to accordeon
 
 #### 0.1.5 (21/05/14)